+++ changelog.md
@@ -6,6 +6,7 @@
 > This version is not released yet and is under active development.
 
 - Upgrade to Click 8.2.0.
+- Add support for custom deprecated messages on commands and parameters.
 - Remove dependency on `pallets-sphinx-themes`.
 - Drop supports for Python 3.10.